CloudFunction/requirements.txt
# Function dependencies, for example:
# package>=version
pandas
pandas-gbq
google-auth